refactor: extract shared landing components to reduce duplication
- Create components/landings/shared/ with reusable LandingFooter,
  LandingNav, LandingWhatsAppBand, LandingPillars, and shared constants
- Replace CorpFooter, CorpNav, CorpWhatsAppBand, CorpPillars with
  thin wrappers around shared components
- Move SOCIAL_LINKS and fadeUp to shared constants (re-exported)
- Use numeric rotateDeg instead of parsing Tailwind class strings
- Eliminates ~250 lines of copy-paste duplication
