+++ components/landings/corporativo/constants.ts
@@ -1,25 +1,12 @@
 import { UsersThree, Sparkle, ClipboardText } from '@phosphor-icons/react';
-import type { Icon } from '@phosphor-icons/react';
+import type { PillarItem } from '../shared/LandingPillars';
+
+export { SOCIAL_LINKS, fadeUp } from '../shared/constants';
 
 export const WHATSAPP_MESSAGE =
     'Oi! Vi que vocês fazem viagens corporativas. Quero entender como funciona pra minha empresa.';
 
-export const SOCIAL_LINKS = {
-    instagram: 'https://instagram.com/anhangaviagens',
-    facebook: 'https://facebook.com/profile.php?id=61585422494271',
-};
-
-export interface Pillar {
-    icon: Icon;
-    title: string;
-    description: string;
-    bg: string;
-    accent: string;
-    iconColor: string;
-    rotate: string;
-}
-
-export const PILLARS: Pillar[] = [
+export const PILLARS: PillarItem[] = [
     {
         icon: UsersThree,
         title: 'Consultor só seu',
@@ -27,7 +14,7 @@ export const PILLARS: Pillar[] = [
         bg: 'bg-blue-100',
         accent: 'border-blue-200',
         iconColor: 'text-blue-600',
-        rotate: '-rotate-1',
+        rotateDeg: -1,
     },
     {
         icon: Sparkle,
@@ -36,7 +23,7 @@ export const PILLARS: Pillar[] = [
         bg: 'bg-emerald-100',
         accent: 'border-emerald-200',
         iconColor: 'text-emerald-600',
-        rotate: 'rotate-2',
+        rotateDeg: 2,
     },
     {
         icon: ClipboardText,
@@ -45,15 +32,6 @@ export const PILLARS: Pillar[] = [
         bg: 'bg-orange-100',
         accent: 'border-orange-200',
         iconColor: 'text-orange-600',
-        rotate: '-rotate-2',
+        rotateDeg: -2,
     },
 ];
-
-export const fadeUp = {
-    hidden: { opacity: 0, y: 28 },
-    visible: (i: number = 0) => ({
-        opacity: 1,
-        y: 0,
-        transition: { delay: i * 0.1, duration: 0.55, ease: 'easeOut' as const },
-    }),
-};
